Break up `ObservationsControllerTest`
This breaks up the 3000+ line test into multiple files.

Rubocop takes a long time to get thru the single file, and it can be hard to find what you're looking for.
